requirements.txt
black==22.8.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.0.3
Flask==2.0.2
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.0.1
Jinja2==3.0.2
Markdown==3.4.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
nanoid==2.0.0
pathspec==0.10.1
platformdirs==2.5.2
psycopg2-binary==2.9.1
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.26
stripe==4.1.0
tomli==2.0.1
urllib3==1.26.12
Werkzeug==2.0.2
